[DELIA-43983] [Thunder]-System Plugin Issues - fix

+++ SystemServices/SystemServices.cpp
@@ -1299,7 +1299,7 @@ namespace WPEFramework {
                 response["downloadPercent"] = m_downloadPercent;
                 retStatus = true;
             } else {
-                response["downloadPercent"] = 0;
+                response["downloadPercent"] = -1;
                 retStatus = true;
             }
             returnResponse(retStatus);
@@ -1643,25 +1643,28 @@ namespace WPEFramework {
 
         /***
          * @brief : To get cashed value .
-         * @param1[in]  : {"params":{"param":{"cacheKey":"<string>"}}}
+         * @param1[in]  : {"params":{"key":"<string>"}}
          * @param2[out] : {"result":{"<cachekey>":"<string>","success":<bool>}}
          * @return      : Core::<StatusCode>
          */
         uint32_t SystemServices::getCachedValue(const JsonObject& parameters,
-                JsonObject& response)
-        {
-            JsonObject param;
-            param.FromString(parameters["param"].String());
-            std::string key = param["cacheKey"].String();
-            LOGWARN("key: %s\n", key.c_str());
-            response[(key.c_str())] = m_cacheService.getValue(key).String();
-            returnResponse(true);
-        }
+			JsonObject& response)
+	{
+		bool retStat = false;
+		std::string key = parameters["key"].String();
+		LOGWARN("key: %s\n", key.c_str());
+		if (key.length()) {
+			response[(key.c_str())] = m_cacheService.getValue(key).String();
+			retStat = true;
+		} else {
+			populateResponseWithError(SysSrv_UnSupportedFormat, response);
+		}
+		returnResponse(retStat);
+	}
 
         /***
          * @brief : To set cache value.
-         * @param1[in]  : {"params":{"param":{"cacheKey":"<string>",
-         *                 "cacheValue":<double>}}}
+         * @param1[in]  : {"params":{"key":"<string>","value":<double>}}
          * @param2[out] : {"jsonrpc":"2.0","id":3,"result":{"success":<bool>}}
          * @return      : Core::<StatusCode>
          */
@@ -1669,59 +1672,68 @@ namespace WPEFramework {
                 JsonObject& response)
         {
             bool retStat = false;
-            JsonObject param;
-            param.FromString(parameters["param"].String());
-            std::string key = param["cacheKey"].String();
-            std::string value = param["cacheValue"].String();
+            std::string key = parameters["key"].String();
+            std::string value = parameters["value"].String();
 
-            if (m_cacheService.setValue(key, value)) {
-                retStat = true;
-            } else {
-                LOGERR("Accessing m_cacheService.setValue failed\n.");
-            }
-            returnResponse(retStat);
+	    if (key.length() && value.length()) {
+		    if (m_cacheService.setValue(key, value)) {
+			    retStat = true;
+		    } else {
+			    LOGERR("Accessing m_cacheService.setValue failed\n.");
+			    populateResponseWithError(SysSrv_Unexpected, response);
+		    }
+	    } else {
+		    populateResponseWithError(SysSrv_UnSupportedFormat, response);
+	    }
+	    returnResponse(retStat);
         }
 
         /***
          * @brief : To check if key value present in cache.
-         * @param1[in]  : {"params":{"param":{"cacheKey":"<string>"}}}
+         * @param1[in]  : {"params":{"key":"<string>"}}
          * @param2[out] : {"jsonrpc":"2.0","id":3,"result":{"success":<bool>}}
          * @return      : Core::<StatusCode>
          */
         uint32_t SystemServices::cacheContains(const JsonObject& parameters,
                 JsonObject& response)
         {
-            bool retStat = false;
-            JsonObject param;
-            param.FromString(parameters["param"].String());
-            std::string key = param["cacheKey"].String();
-            if (m_cacheService.contains(key)) {
-                retStat = true;
-            } else {
-                LOGERR("Accessing m_cacheService.contains failed\n.");
-            }
-            returnResponse(retStat);
+		bool retStat = false;
+		std::string key = parameters["key"].String();
+		if (key.length()) {
+			if (m_cacheService.contains(key)) {
+				retStat = true;
+			} else {
+				LOGERR("Accessing m_cacheService.contains failed\n.");
+				populateResponseWithError(SysSrv_Unexpected, response);
+			}
+		} else {
+			populateResponseWithError(SysSrv_UnSupportedFormat, response);
+		}
+		returnResponse(retStat);
         }
 
         /***
          * @brief : To delete the key value present in cache.
-         * @param1[in]  : {"params":{"param":{"cacheKey":"<string>"}}}
+         * @param1[in]  : {"params":{"key":"<string>"}}
          * @param2[out] : {"jsonrpc":"2.0","id":3,"result":{"success":<bool>}}
          * @return      : Core::<StatusCode>
          */
         uint32_t SystemServices::removeCacheKey(const JsonObject& parameters,
                 JsonObject& response)
         {
-            bool retStat = false;
-            JsonObject param;
-            param.FromString(parameters["param"].String());
-            std::string key = param["cacheKey"].String();
-            if (m_cacheService.remove(key)) {
-                retStat = true;
-            } else {
-                LOGERR("Accessing m_cacheService.remove failed\n.");
-            }
-            returnResponse(retStat);
+		bool retStat = false;
+		std::string key = parameters["key"].String();
+		if (key.length()) {
+			if (m_cacheService.remove(key)) {
+				retStat = true;
+			} else {
+				LOGERR("Accessing m_cacheService.remove failed\n.");
+				populateResponseWithError(SysSrv_Unexpected, response);
+			}
+		} else {
+			populateResponseWithError(SysSrv_UnSupportedFormat, response);
+		}
+		returnResponse(retStat);
         }
 
         /***
@@ -1790,28 +1802,30 @@ namespace WPEFramework {
          */
         uint32_t SystemServices::getLastDeepSleepReason(const JsonObject& parameters,
                 JsonObject& response)
-        {
-            bool retAPIStatus = false;
-            string reason;
+	{
+		bool retAPIStatus = false;
+		string reason;
 
-            if (Utils::fileExists(STANDBY_REASON_FILE)) {
-                    std::ifstream inFile(STANDBY_REASON_FILE);
-                    if (inFile) {
-                        std::getline(inFile, reason);
-                        inFile.close();
-                        retAPIStatus = true;
-                    } else {
-                        populateResponseWithError(SysSrv_FileAccessFailed, response);
-                    }
-                } else {
-                    populateResponseWithError(SysSrv_FileNotPresent, response);
-                }
+		if (Utils::fileExists(STANDBY_REASON_FILE)) {
+			std::ifstream inFile(STANDBY_REASON_FILE);
+			if (inFile) {
+				std::getline(inFile, reason);
+				inFile.close();
+				retAPIStatus = true;
+			} else {
+				populateResponseWithError(SysSrv_FileAccessFailed, response);
+			}
+		} else {
+			populateResponseWithError(SysSrv_FileNotPresent, response);
+		}
 
-                if (retAPIStatus && reason.length()) {
-                    response["lastDeepSleepReason"] = reason;
-            }
-            returnResponse(retAPIStatus);
-        }
+		if (retAPIStatus && reason.length()) {
+			response["reason"] = reason;
+		} else {
+			response["reason"] = "";
+		}
+		returnResponse(retAPIStatus);
+	}
 
         /***
          * @brief : Used to clear last deep sleep reason.
@@ -2079,24 +2093,24 @@ namespace WPEFramework {
 
         /***
          * @brief : Enables XRE Connection Retension option.
-         * @param1[in]  : {"params":{"param":<bool>}}
+         * @param1[in]  : {"params":{"enable":<bool>}}
          * @param2[out] : "result":{"success":<bool>}
          * @return      : Core::<StatusCode>
          */
         uint32_t SystemServices::enableXREConnectionRetention(const JsonObject& parameters,
                 JsonObject& response)
-        {
-            bool enable = false, retstatus = false;
-            int status = SysSrv_Unexpected;
+	{
+		bool enable = false, retstatus = false;
+		int status = SysSrv_Unexpected;
 
-                enable = parameters["param"].Boolean();
-                if ((status = enableXREConnectionRetentionHelper(enable)) == SysSrv_OK) {
-                    retstatus = true;
-                } else {
-                    populateResponseWithError(status, response);
-            }
-            returnResponse(retstatus);
-        }
+		enable = parameters["enable"].Boolean();
+		if ((status = enableXREConnectionRetentionHelper(enable)) == SysSrv_OK) {
+			retstatus = true;
+		} else {
+			populateResponseWithError(status, response);
+		}
+		returnResponse(retstatus);
+	}
 
         /***
          * @brief : collect device state info.
@@ -2243,59 +2257,45 @@ namespace WPEFramework {
          */
         uint32_t SystemServices::setDevicePowerState(const JsonObject& parameters,
                 JsonObject& response)
-        {
-            bool retVal = false;
-            string sleepMode;
-            ofstream outfile;
-            outfile.open(STANDBY_REASON_FILE, ios::out);
-                JsonObject paramIn, paramOut;
-                string state = parameters["powerState"].String();
-                string reason = parameters["standbyReason"].String();
+	{
+		bool retVal = false;
+		string sleepMode;
+		ofstream outfile;
+		JsonObject paramIn, paramOut;
+		string state = parameters["powerState"].String();
+		string reason = parameters["standbyReason"].String();
+		/* Power state defaults standbyReason is "application". */
+		reason = ((reason.length()) ? reason : "application");
 
-
-                if(state=="STANDBY")
-                {
-
-
-                    if (SystemServices::_instance) {
-                        SystemServices::_instance->getPreferredStandbyMode(paramIn, paramOut);
-                        /* TODO: parse abd get the sleepMode from paramOut */
-                        sleepMode= paramOut["preferredStandbyMode"].String();
-
-
-                    LOGWARN("Output of preferredStandbyMode: '%s'", sleepMode.c_str());
-
-                    }
-                    else {
-                        LOGWARN("SystemServices::_instance is NULL.\n");
-                    }
-                    if(convert("DEEP_SLEEP",sleepMode))
-                    {
-
-                        retVal = CPowerState::instance()->setPowerState(sleepMode);
-                    }
-                    else{
-
-                        retVal = CPowerState::instance()->setPowerState(state);
-                    }
-
-
-                    if (outfile) {
-                        outfile << reason;
-                        outfile.close();
-
-                    }
-                    else {
-                        printf(" GZ_FILE_ERROR: Can't open file for write mode\n");
-                    }
-
-                }
-                else {
-                    retVal = CPowerState::instance()->setPowerState(state);
-                    LOGERR("this platform has no API System and/or Powerstate\n");
-            }
-            returnResponse(retVal);
-        }//end of setPower State
+		if (state == "STANDBY") {
+			if (SystemServices::_instance) {
+				SystemServices::_instance->getPreferredStandbyMode(paramIn, paramOut);
+				/* TODO: parse abd get the sleepMode from paramOut */
+				sleepMode= paramOut["preferredStandbyMode"].String();
+				LOGWARN("Output of preferredStandbyMode: '%s'", sleepMode.c_str());
+			} else {
+				LOGWARN("SystemServices::_instance is NULL.\n");
+			}
+			if (convert("DEEP_SLEEP", sleepMode)) {
+				retVal = CPowerState::instance()->setPowerState(sleepMode);
+			} else {
+				retVal = CPowerState::instance()->setPowerState(state);
+			}
+			if (outfile) {
+				outfile.open(STANDBY_REASON_FILE, ios::out);
+				if (outfile.is_open()) {
+					outfile << reason;
+					outfile.close();
+				}
+			} else {
+				printf(" GZ_FILE_ERROR: Can't open file for write mode\n");
+			}
+		} else {
+			retVal = CPowerState::instance()->setPowerState(state);
+			LOGERR("this platform has no API System and/or Powerstate\n");
+		}
+		returnResponse(retVal);
+	}//end of setPower State
 #endif /* HAS_API_SYSTEM && HAS_API_POWERSTATE */
 
         /***
